+++ CHANGES.md
@@ -1,5 +1,6 @@
 0.6.1
 -----
+* Celluloid#links obtains Celluloid::Links for a given actor
 * The #class method is now proxied to the actor
 
 0.6.0